featureimage/IPMPROG-2969-metric-storage

+++ src/persistance.h
@@ -30,11 +30,11 @@ class Row;
 } // namespace tntdb
 
 // t_bios_measurement/value
-typedef int32_t m_msrmnt_value_t;
+typedef int64_t m_msrmnt_value_t;
 // t_bios_measurement/scale
 typedef int16_t m_msrmnt_scale_t;
 // t_bios_measurement_topic/id
-typedef uint16_t m_msrmnt_tpc_id_t;
+typedef uint32_t m_msrmnt_tpc_id_t;
 // t_bios_discovered_device/id_discovered_device
 typedef uint16_t m_dvc_id_t;
